fix: moving categoryPage and style
