+++ client/src/styles/Footer.scss
